Do a bit more work in vm-initialization and a bit less on every run.
It probably isn't actually a time savings, but I like the precedent it
sets of only doing per-request work in render(), and all the
per-component initialization in getVMContext().  Hopefully this will
help us divvy things up properly when we add aphrodite support.

I wanted to assign Component in getVMContext as well, but I couldn't
figure out how to make it available to render() since the
implementation of runInContext purposefully puts each block of code in
its own scope (so vars don't leak from one invocatation to the
other).  We could change that, but I don't think it's worth it for
this.

Auditors: jlfwong

Test Plan:
npm test