Remove the need to depend on canvas.
Canvas was only used in qTip2, and qTip2 even has a check for whether
canvas is installed (and works correctly when it's not).  But that
check was broken for us because of the way that jsdom mocks out the
canvas.

We mock it out a different way, and now we can require qTip2 just fine
without needing a canvas!

Pair programmed with jlfwong

Auditors: jlfwong

Test Plan:
npm test