Update the benchmark script to talk to prod as well as dev.
Summary:
I had to refactor a bit of stuff since the information is presented a
bit differently in prod and dev.  But it wasn't too bad.  The code is
a bit better factored now anyway.  But still ugly amounts of
indentation!

Use the promise of promises to get rid of a bunch of indentation.

The file was substantially rewritten, but the test still works!

Test Plan:
I changed the 4th arg to render() to be
   https://www.khanacademy.org
instead of
   localhost:8080
and ran this script, and it said:
   javascript/content-library-package/components/concept-thumbnail.jsx: 542
same as when I ran it in dev.

(same)

Reviewers: jlfwong

Reviewed By: jlfwong

Subscribers: emily, john

Differential Revision: https://phabricator.khanacademy.org/D24588