Pass the secret in when doing the /render request!
Otherwise, we won't be able to run the benchmark on a prod
react-render-server.

Auditors: jlfwong

Test Plan:
We don't have a prod server running yet, so I put in
   console.log(JSON.stringify(reqBody));
right after reqBody was defined, and then ran
    npm run serve_local
    tools/benchmark_ka_component.js --dev-render

It printed out the post data with the secret embedded, and the
benchmark run completed successfully, meaning that the post data
passed the /render sanity-checking.