Move to winston for logging, and use logging instead of console.
Summary:
It's unclear to me how console output makes its way into the VM server
logs, and it's probably better to use a real logging library anyway.
Now we use winston, as suggested at
   https://cloud.google.com/nodejs/getting-started/logging-application-events

I set things up the way they suggest on that page.  I also changed all
the console.log to logging.*.

Finally, I think part of the difficulty of debugging a previous error
with the secrets file is that it did a process.exit(), which maybe
causes the request not to get logged at all?  Now I just throw, which
returns a 'normal' 500, making it easier to debug.

Finally, I had to get rid of the --use-strict flag for node because
winston uses some non-strict features (in particular, octal literals).
It never seemed to work right anyway.  While in the area, I moved some
'use strict's down below the top-of-file docstring to make the
docstring easier to read.

Test Plan:
npm test

Also, ran
   npm run serve_local
and then
   tools/benchmark_ka_component.js --dev-render
and had it complete without error.

Reviewers: jlfwong

Reviewed By: jlfwong

Subscribers: emily, john

Differential Revision: https://phabricator.khanacademy.org/D24635