Throw, rather than die, when we can't read the secrets file.
There was a bug in this code that was very difficult to debug, I think
maybe because it caused the server to die.  Now we just make the
server return a 500, which hopefully will be easier for us to debug in
the future!

Auditors: jlfwong

Test plan:
In main.js, I temporarily commented out the line
    renderSecret.matches = (actual) => true;
In benchmark_ka_component.js, I temporarily commented out the line
    secret: secret.get(),

I then changed secret.js to change 'trim' to 'trimmm'.
I then ran
   npm run serve_local
   curl -D- -H "Content-type: application/json" -d '{"secret":"'`cat ../secret`'", "urls":["https://www.khanacademy.org/genfiles/javascript/en/corelibs-package-31375e.js","https://www.khanacademy.org/genfiles/javascript/en/corelibs-legacy-package-fbfab0.js","https://www.khanacademy.org/genfiles/javascript/en/shared-package-1e468a.js","https://www.khanacademy.org/genfiles/javascript/en/shared-styles-package-32d405.js","https://www.khanacademy.org/genfiles/javascript/en/hover-card-package-de8d87.js","https://www.khanacademy.org/genfiles/javascript/en/react-package-9b7fb9.js","https://www.khanacademy.org/genfiles/javascript/en/react-components-package-a7e18c.js","https://www.khanacademy.org/genfiles/javascript/en/flux-package-838d8f.js","https://www.khanacademy.org/genfiles/javascript/en/tasks-package-43bf44.js","https://www.khanacademy.org/genfiles/javascript/en/tutorial-package-8e5302.js","https://www.khanacademy.org/genfiles/javascript/en/video-package-9b200b.js","https://www.khanacademy.org/genfiles/javascript/en/content-library-package-9ac69b.js"],"path":"./javascript/content-library-package/components/concept-thumbnail.jsx","props":{"domain":"math","kind":"concept","progressData":{"skillsCompleted":3,"skillsTotal":32},"url":"/images/topic-icons-large/linear_equations.png"}}' 127.0.0.1:8060/render

I got a 500 response, with the text of the page being the 'TypeError'
message, and the server logged a 500 (plus the error text) and did not
die.