Add an 'in flight' cache so we only fetch a given resource once.
Before this change, if 20 requests came in at the same time, each
wanting the same package, we'd fetch that package from webapp 20
times.  Now we only fetch it once, reusing the fetch-promise while the
fetch is in flight (after the fetch finishes, the response will be in
the cache so we don't need the promise at all anymore).

Auditors: jlfwong

Test Plan:
npm test

I also commented out the line
//       inFlightRequests[inFlightCacheKey] = retval;
and re-ran the tests and saw that my new "should only test once" test
failed.