Log information about factors that might relate to how long a /render takes.
Summary:
We log:
1) The name of the component being rendered (but not the size/contents
   of its props!)
2) How many concurrent requests are in flight at the same time as this
   one
3) How many package-fetches we needed to do for this request.

The goal is to download this logs data -- perhaps via the method in
DEBUGGING -- and then analyze these numbers to come up with a function
that estimates how long it will take to serve a request as a function
of these variables -- hopefully only (2) and (3).

We'll use that estimation function to early-bail on requests that we
"know" will take longer than the client is willing to wait.  This may
also serve as a useful input to the google auto-scaler, one day.

I made the profile API a bit more flexible and also maybe simpler:
you can pass in an end message different from the start one, and if
you don't want a start message at all, just leave out that arg (rather
than having to pass in a message and then a second `false` arg).

While researching this, I discovered we didn't need our own
error-handling middleware; express has a default which is just fine
for us.  So I took our custom one out.

Resolve the TODO to log how many package-fetches a given request incurred.

It was annoying to thread this data through.  The fetchPackage API is
a lot messier now. :-(  On the plus side, we get to test some of the
'was this cached?' logic more directly now.

Only log render-stats for 2xx response.

For errors, the stats will only confuse matters.

Test Plan:
I ran
   npm run debug
   tools/benchmark_ka_component.js --dev-render
and saw it emit a bunch of PROFILE lines, ending with:
   debug: PROFILE: /render for ./javascript/content-library-package/components/concept-thumbnail.jsx: had 0 concurrent in-flight requests (finished after 669.1947 ms)

npm test

I also ran
   npm run debug
   tools/benchmark_ka_component.js --dev-render
and saw the last profile line say:
    debug: PROFILE: render-stats for ./javascript/content-library-package/components/concept-thumbnail.jsx: {"pendingRenderRequests":0,"packageFetches":13} (finished after 717.2954 ms)
which seems right.

npm run debug

I then did
   tools/bennent.js --dev-render
and saw that the render-stats line was still printed.  I then did
   curl -H "Content-type: application/json" -d '{}' 127.0.0.1:8060/render
(which gives a 400 error) and saw that it wasn't.

Reviewers: jlfwong

Reviewed By: jlfwong

Subscribers: emily, john

Differential Revision: https://phabricator.khanacademy.org/D24746