Add periodic reporting of queries per second, and simplify some code.
Summary:
I reworked f9a9d7b in a much simpler way, by keeping the old (bad)
order of the for-loop used to create the render-fetches, but then
shuffling the list of fetches.  In other words, we create the list of
things to fetch before doing any fetches, and then shuffle in the
middle.  A lot more readable now, I think.

I also simplified (at least to me) the throttling code a bit.  Now we
have max-concurrent-request 'render loops', each of which just renders
repeatedly, taking the next thing-to-render from a global queue.  That
way, we always have max-concurrent-request renders going on at once!

Test Plan:
   % tools/benchmark_ka_component.js -n2 -r1 ../webapp/javascript/editor-package/search-result-preview-widget.jsx.fixture.js ../webapp/javascript/content-library-package/components/concept-thumbnail.jsx.fixture.js ../webapp/javascript/learn-storm-package/kickoff-pledge.jsx ../webapp/javascript/learn-storm-package/kickoff-pledge.jsx.fixture.js
    Skipping ../webapp/javascript/learn-storm-package/kickoff-pledge.jsx: cannot infer component from /home/csilvers/khan/webapp/javascript/learn-storm-package/kickoff-pledge.jsx
    javascript/learn-storm-package/kickoff-pledge.jsx: 1393 191ms
    REQUESTS IN THE LAST SECOND: 2
    javascript/content-library-package/components/concept-thumbnail.jsx: 1679 194ms
    javascript/learn-storm-package/kickoff-pledge.jsx: 1395 193ms
    Skipping /home/csilvers/khan/webapp/javascript/editor-package/search-result-preview-widget.jsx.fixture.js: Error: Cannot find module 'underscore'
    javascript/content-library-package/components/concept-thumbnail.jsx: 1679 194ms
    Skipping /home/csilvers/khan/webapp/javascript/editor-package/search-result-preview-widget.jsx.fixture.js: Error: Cannot find module 'underscore'
    DONE!

The QPS stuff is being logged (only once since this is running pretty
fast).  The renders seem to be shuffled randomly -- hard to tell with
only 4 renders, but they are at least in a different order than the
for-loops would have put them in.  We still skip the preview-widget
twice, but still only print the 'cannot infer component' once, so that
functionality hasn't regressed.

I then ran
   tools/benchmark_ka_component.js ~/khan/webapp/javascript/*-package/*.fixture.* -r1
and it emitted the QPS logline many times, usually 3-6 QPS each time
(which makes sense given the -r1), and otherwise seemed to give the
correct output.

Reviewers: jlfwong

Reviewed By: jlfwong

Differential Revision: https://phabricator.khanacademy.org/D24786