Move more work out of the main render() loop.
I am suspecting that the actual constraint on our load-testing system
is the driver (this script), rather than the server.  To help with
that, I'm moving as much work as possible out of the request-loop and
into the function that creates the render-args.  While this makes only
minimal difference when running with -n1 (the default), it should help
a lot with larger -n values.

The render() function args become a lot simpler too!

Auditors: jlfwong

Test Plan:
   tools/benchmark_ka_component.js -n2 -r1 ../webapp/javascript/editor-package/search-result-preview-widget.jsx.fixture.js ../webapp/javascript/content-library-package/components/concept-thumbnail.jsx.fixture.js ../webapp/javascript/learn-storm-package/kickoff-pledge.jsx ../webapp/javascript/learn-storm-package/kickoff-pledge.jsx.fixture.js
shows the same output as in previous runs.