Add a simple script to validate that a component can be server-side rendered.
This is intended for use with webapp's
react_util/server_side_render_test.py.

Auditors: jlfwong

Test Plan:
In webapp, ran
   tools/runtests.py --max-size=large react_util/server_side_render_test.py
It completed successfully.

I then changed the cache policy in validate_components.js from
'ignore' to 'yes', and ran the test again, and saw it fail as so:

    AssertionError: ./javascript/page-package/stateful-header.jsx #0: ERROR: TypeError: Cannot read property 'get' of undefined
    ./javascript/content-library-package/curation-page.jsx #0: ERROR: TypeError: Cannot read property 'get' of undefined
    ./javascript/content-library-package/curation-page.jsx #1: ERROR: TypeError: Cannot read property 'get' of undefined
    ./javascript/content-library-package/curation-page.jsx #2: ERROR: TypeError: Cannot read property 'get' of undefined
    ./javascript/content-library-package/curation-page.jsx #3: ERROR: TypeError: Cannot read property 'get' of undefined
    ./javascript/style-package/style-guide.jsx #0: ERROR: TypeError: Cannot read property 'get' of undefined
    ./javascript/about-package/alumni.jsx #0: ERROR: TypeError: Cannot read property 'get' of undefined

This shows that it is properly iterating over all the relevant
components and all their relevant fixtures.  (The 'get' error is
because I never initialize the cache, so trying to use the cache
fails.)