Add .travis.yml file to integrate with Travis CI
I noticed some of the tests were failing because we forgot to update some of our
stubs when we integrated with Graphite, and I think we would've been more likely
to notice if we had Travis CI running the tests on commits.

We can separately add a review hook for Arcanist to enable running tests with
`arc unit`, but this is easier for now, or we could use something like
https://github.com/zerodiff/traphic to make Phabricator talk to Travis CI

Test Plan:
Cross fingers and check https://travis-ci.org/Khan/react-render-server after
this is pushed.

Auditors: csilvers