Increase the limit on the request body size to 5 megabytes
Summary:
When loading up a tutorial page with a lot of content, we were seeing failures
because the request payload was too large. It turns out that this limitation is
enforced by express body-parser.

So we'll just bump it here from 100kb (the default) to 5mb.

Test Plan:
Cross fingers and load
https://www.khanacademy.org/math/cc-fifth-grade-math/cc-5th-arith-operations/cc-5th-add-sub-decimals/v/adding-decimals-example-1
with SSR turned on after this is deployed.

Auditors: csilvers