Stop all old versions, even if we don't delete them.
Summary:
This way we won't get charged for them.

I also took the opportunity to change from 'app modules' to 'app
services', the sparkly new name.

Test Plan:
I ran the latter part of the script manually on my machine -- starting
with assigning VERSIONS -- without error.  It said:
   Deleting old version 160302-1216-2b6ae5c0c469
   Stopping old version 160322-1701-09a7b562d2b0
   Stopping old version 160321-1101-d82863c5657c
   Stopping old version 160317-1437-1547da1dd0da
   Stopping old version 160308-1355-ca3a43c861d7

For the last 4 it also gave back a 400 saying that the version was
already stopped, which is expected since we had manually stopped
them.

Reviewers: benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D27208