Deep copy incoming props before rendering them.
Summary:
I was hitting an error in the prod react-render-server where it
was complaining that `.includes` on a certain array wasn't a function.
We [explicitly polyfill
`Array.prototype.includes`](https://www.khanacademy.org/r/js-polyfills),
but it somehow wasn't making it onto this array. The same call was
succeeding on the client, during its initial render.

I traced it back to a problem where the array inside of the props was
being generated in the global context of the react-render-server, which
lacks the polyfills that the inner contexts we use to render have.

This fixes the problem by re-creating the props inside of the render
context, so that the generated arrays have the appropriate polyfills.

Test Plan:
 - `npm run test`
 - Run `node src/main.js --dev --port=8040`
 - Visit
   http://localhost:8081/math/in-sixth-grade-math/knowing-numbers/new-topic-2015-11-17T22:50:42.295Z/e/expressions-with-parentheses?ssrk=2&norehydratek=1
   in webapp
 - See that the exercise component successfully server-side renders, and
   there are no errors in the react-render-server output.

Reviewers: csilvers, john

Reviewed By: john

Differential Revision: https://phabricator.khanacademy.org/D30119