Remove the unhandledRejection handler.
Summary:
It was saying every rejection was unhandled, which may be true -- I'm
not 100% sure how rejection-handling works.  In any case, we seem to
be doing a good job of logging every time we reject, so I don't think
we need to do more logging here.

I like getting rid of this because its default output is very long,
since it includes the entire http response from the page-fetch, which
is typically longer than 8K which is the limit for loglines in
appengine (or at least in the logs viewer).  I'm hoping this will make
the logs more usable.

Test Plan:
I ran
   npm run serve_local
and in a separate terminal did
   curl -H content-type:application/json -d '{"urls":["https://www.khanacademy.org/crash"],"path":"./javascript/server-package/test-component.jsx","props":{},"secret":"sekret"}' localhost:8060/render
and saw it say:
   error: timed out while fetching https://www.khanacademy.org/crash

I also ran
   curl -H content-type:application/json -d '{"urls":["https://www.khanacademy.org/crashh"],"path":"./javascript/server-package/test-component.jsx","props":{},"secret":"sekret"}' localhost:8060/render
and saw it say:
   error: Fetching failure: Error: cannot GET /crashh (404):  Error: Not Found
    at Request.callback (/home/csilvers/khan/react-render-server/node_modules/superagent/lib/node/index.js:795:17)
    at IncomingMessage.<anonymous> (/home/csilvers/khan/react-render-server/node_modules/superagent/lib/node/index.js:997:12)
    at emitNone (events.js:72:20)
    at IncomingMessage.emit (events.js:166:7)
    at endReadableNT (_stream_readable.js:913:12)
    at nextTickCallbackWith2Args (node.js:442:9)
    at process._tickCallback (node.js:356:17)

The stack isn't that useful, it doesn't look like, but we'll keep it
for now.

Reviewers: john, emily

Reviewed By: john, emily

Differential Revision: https://phabricator.khanacademy.org/D30604