Add support for Apollo to React Render Server.
Summary:
This diff adds support for Apollo to React Render Server. In short: Apollo is a way for us to define our data requirements for a React component on the component itself, rather than living somewhere in Python-land. Each component can say what data is needed and then all of those requirements are batched together into a single request to our new GraphQL API endpoint. This is a highly efficient way of collecting the data and feeding it in to the components.

Since the logic for determining the data requirements is on the components themselves we will need to figure out the requirements in React Render Server (using `react-apollo`), make a request to the server, and then feed the data into the components before rendering them. This will result in a pre-rendered version of our components that already have the data in them (along with a JSON blob that we can insert into the page, as well).

I have a follow-up diff coming that adds support for this to our render_react template tag.

Test Plan:
I ran `npm test` and it passed.

Additionally, within webapp, I visited a page that I had written that uses this technology: http://localhost:8080/coach/review and confirmed that the page was server-side rendered. Specifically, the page said "Test Class 1" (which could only have come from a successful API request) and the JSON data from the GraphQL APi was embedded in the page. Finally, watching the network console I confirmed that no network requests were made to the GraphQL API, all the data that was needed was in the page.

Reviewers: briangenisio, kevinb, emily

Reviewed By: briangenisio, kevinb, emily

Subscribers: dhruv, csilvers

Differential Revision: https://phabricator.khanacademy.org/D33626