Bump the memory request up
In old-flex (`vm: true`), we had to use predefined machine sizes, so even
though we requested 1GB of memory we actually got an `n1-standard-1` with
3.75GB of memory.  In new-flex (`env: flex`) we can have custom machine sizes,
which means we actually get what we ask for (plus 0.4GB of expected overhead).
So now we were getting a lot less, which it appears, empirically, is not
enough -- see https://app.asana.com/0/265475162757451/318889581592147.  This
commit changes us to request about what we did before, which we seem to be
using the bulk of.  I don't really understand why we use so much, so I left a
TODO for that.

Test plan: deploy it, don't see a spike in react-render errors.

Auditors: csilvers