Use a subprocess to do rendering, and provide a render-timeout (default 1s).
Summary:
Some render requests take a long time to execute.  Since they're
CPU-bound, node has no good way of killing them before they finish.
So we now run them in a subprocess, and just kill the subprocess if
it's taking too long.  We use a handy library that does this for us.

We default to running as many workers as there are cores on our
machine, and letting each worker run one task at a time (these are the
handy-library defaults).  Future requests will queue up, so it's still
possible that when a request takes a long time, other requests will
take a long time too (due to waiting), even if they themselves don't
require much work.

One downside of this approach is that our timeout is only for time
spent //rendering//.  So a request can take as long as the fetch
timeout + the render timeout before we time it out, which isn't ideal.
Plus, the time that requests spend waiting for a worker to open up, so
they can do their rendering, isn't counted in any timeout that we have.
There may or may not be ways to improve this situation, possibly just
by upping the number of workers we have.

Helps resolve https://app.asana.com/0/31965416896056/323826299198425

Upgradee to node v8.

worker-nodes.js required at least node v6.6.0, and we might as well go
all the way.

A test started failing, presumably due to the way node v8 serializes
errors that is different from how node v4 does.  I just catch that
error explicitly and do my own serialization.  Now we have a better
error message in any case.

(Hackily) add proper cache support.

The tricky bit here is that there's no way to initialize the cache in
each of the workers, and in general no way to tell the workers what
cache size we want.  We //could// make the cache size an argument to
the `render()` call, but that's a pretty serious violation of
responsibilities.

Instead I do something almost as bad, and make a temporary-file to be
the entrypoint of the workers.  This temporary file initializes the
cache and then returns the render() function.

It all works except when I run tests, I get this error after the tests
successfully pass:
```
Error: Cannot find module '/home/csilvers/khan/react-render-server/src/render_cache_10000_QblMeU.js'
    at Function.Module._resolveFilename (module.js:489:15)
    at Function.Module._load (module.js:439:25)
    at Module.require (module.js:517:17)
    at require (internal/module.js:11:18)
    at setupModule (/home/csilvers/khan/react-render-server/node_modules/worker-nodes/lib/worker/child-loader.js:12:15)
    at process.<anonymous> (/home/csilvers/khan/react-render-server/node_modules/worker-nodes/lib/worker/child-loader.js:55:20)
    at emitTwo (events.js:125:13)
    at process.emit (events.js:213:7)
    at emit (internal/child_process.js:774:12)
    at _combinedTickCallback (internal/process/next_tick.js:141:11)
```

It looks like the worker-node library is trying to start a new worker
after the tempfile code has cleaned up all the tempfiles.  I don't
know why this is happening: I'm calling `renderWorkers.terminate()` in
the teardown of the relevant tests, so you'd think that it wouldn't be
trying to create new instances anymore.  I can't figure it out -- the
stacktrace isn't helpful -- and it's not actually affecting test
correctness, so I'm leaving it be even though it's confusing and
ugly.  Any suggestions welcome!

Use `...args` instead of repeating arguments.

Now that we've updated to node v8, this works just fine.

Add package-lock.json.

I hear this is good for node v8 users, and neutral for people using it
earlier.

Use lazyStart to minimize the weird error message printed after tests.

I figured out what causes the errors that worker-nodes prints after
the tests are complete.  What happens is worker-nodes wants to start,
say, 8 workers.  It is happily starting up workers when the test
finishes (because it's fast).  But it will continue starting up
workers even after we terminate!  This is a bug in the worker-nodes
code.

I fixed this by doing `lazyStart`, which means it will only load up
one worker since we have only one task for it to do.

This works for everything except for the test with the 1ms timeout,
which still fails before the worker has started up.

I'll try to get a fix for this upstream.

Test Plan:
npm test

npm test

(same)

(same)

(same)

(same)

Reviewers: john, emily

Reviewed By: john

Subscribers: benkraft

Differential Revision: https://phabricator.khanacademy.org/D38583