Up the timeout to 4seconds for the /render tests.
Every time we call `/render` in a test we have to spawn a new
subprocess to handle the rendering.  That can take a while; or at
least, apparently it does on Jenkins.  2 seconds is not reliably
enough for these tests, so I double the timeout.

Hopefully the actual time-taken on the real server will be faster than
before, not slower!

Auditors: john

Test plan:
Made these changes on jenkins and ran the test there and saw this:
```
    ✓ should render a simple react component (2024ms)
    ✓ should fail on invalid inputs (49ms)
    ✓ should log render-stats (2022ms)
    ...
```
so it looks like the longer timeout is necessary.