Re-instate worker-nodes, but do all of /render in the subprocess.
Summary:
Re-instate worker-nodes, but do all of /render in the subprocess.

Before, using worker-nodes was too slow because it took too long to
pass all the data to the workers and back.  This is an attempt to
ameliorate that problem by only passing a small(ish) amount of data:
the urls of the deps, the props, etc.  It will still be dozens of
kilobytes, but the old scheme was megabytes and that was a problem.

Finish fixing server-tests in the new subprocess world.

The server-tests want to mock out logs and such, which doesn't work
across subprocess boundaries.  So I split all those tests in two: one
test that makes sure we properly return an error in various
situations, and one test to make sure that we properly log in those
situations.  For the first test, we mimic the prod environment, but
for the second we turn off the functionality that uses the
subprocess.

Since the second test is only testing logging, that should be good
enough to test all the functionality in a way that mirrors what we see
in production.

To get this to work, I had to add yet another envvar. :-( Testing this
new code is definitely messier than I thought it would be.

Speed up startup, and fix an unhandled promise.

Test Plan: npm test

Reviewers: john

Reviewed By: john

Differential Revision: https://phabricator.khanacademy.org/D38818