Do priming during deploy
Summary:
In a recent deploy a bunch of requests timed out because the
number of instances was set to 8 upon set default. It took
about 20 minutes to scale up to a reasonable number of instances
(currently 93).

This change should makes the deploy job do some priming
in order to get the autoscaler to start up a bunch of instances.

I've also reduced the cpu target utilization a little bit for two
reasons: I hope that it will help the priming work better, and
because requests being blocked seems to be associated with
high cpu usage -- so perhaps targeting lower CPU load will
help.

I should also note that I investigated using the admin api to
dynamically set min_total_instances but I wasn't able to make
it work. https://app.asana.com/0/31965416896056/413723877242760
If we are able to make this work in the future, we'll need to
call the api again to lower the min instances on the old version
that we keep running.

Test Plan:
Tested priming endpoint locally to make sure it works.

Fingers crossed. Deploy and see how it works.

Reviewers: csilvers, drosile

Reviewed By: csilvers

Subscribers: benkraft

Differential Revision: https://phabricator.khanacademy.org/D39950