Increase the number of render workers.
Summary:
In interactive testing I've seen that rrs slows
way down when handling a large number of concurrent
request. It will process 100 requests much faster
if there are only a couple concurrent requests than
if there are a bunch of concurrent requests.

One possible reason for this is that we have only
had two render workers per instance. That means
that only two requests can be rendered concurrently.
However the rendering process fetches URLs of javascript
dependencies and may also do IO when calling graphql
queries. In the worst-case each instance will only be
able to handle two requests/second.

Another possible reason (which this diff doesn't
address) is that it's the http server component that's
having issues with concurrency, not the rendering
part. If this change doesn't fix things I'll spend
some time looking into this possibility.

This change increases the number of render workers
to 10 per instance.

This is hard to test without deploying. I was able to
change the number of render workers on a live non-default
instance, but it's hard to do good tests on it, because
the auto-scaler has a tendancy to spin up new instances
when I send a bunch of requests to my hacked instance.

So I'm going to deploy this to be able to test it.

Test Plan:
Mostly will test by deploying and then splitting traffic to
see if this version can handle load.

Reviewers: csilvers, drosile

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D40082