Run server behind nginx
Summary:
Attempts to solve our issues with requests that block
other requests by running the server one request at a
time.

We use nginx as a reverse proxy. The max_conns option
is used to make sure that each server only handles one
request at a time. Unfortunately this option isn't
available except on more recent versions on nginx
so the Dockerfile needs to go to some work to get
a recent version of nginx installed.

The pm2 process manager is used to launch all the node
processes and nginx. It makes sure that the the processes
are respawned if they die.

The flex runtime is changed to custom so that it will
work with the docker image rather than using the stock
node image.

The deploy script is changed to build the docker image
and deploy via the docker image.

Status: WIP, but I think that it's far enough along to
test on prod.

TODO:
- test deploy script

- test in prod and see if this approach actually fixes
  the problem of one slow request slowing other requests.

- ulimit? Not sure if we need to set this the way it
  was previously being set

- update docs

- stderr/stdout logs - I don't think that the node server
  logs are going to be exposed correctly in prod.

Test Plan:
Test docker image locally:
    docker build . -t rrs
    docker run -d -p 8080:8080 rrs
    docker exec -ti CONTAINER /bin/bash

From within the container see that 10 node servers are
running and that nginx is running. Try killing a node
server and see that it re-spawned

From outside the container sending requests to http://localhost:8080

Reviewers: csilvers, drosile

Reviewed By: drosile

Differential Revision: https://phabricator.khanacademy.org/D40165