Improve react render server configuration.
Summary:
Increase target CPU usage. Now that we are making sure that
node won't be handling more than one request at a time, there's
less need to set the target CPU usage lower. Also perhaps
this will help reduce the total number of instances needed
to handle our traffic.

Added a timeout to the nginx proxy. I'm not sure if this will
help much or not. If the long response is due to something that's
causing node to block, then I don't think this will help node.

Added a nginx configuration change to stop it from retrying
requests in case of timeout or failure. We don't want to retry
requests in general because they are unlikely to work better
on another node server, and moreover, there probably isn't
enough time to retry the request.

Added a script to build nginx and pm2 conf files. This script
is run in a build stage. This way I can use python to create
the conf files without having to have python installed in the
final docker image.

Also increased the number of node workers to 15. This is based
on examining memory usage in instances with 10 workers running
and observing that there was substantial memory unused.

Test Plan:
- Tested configuration file generation locally. Examined
  built docker image to confirm that conf files were correctly
  generated.

- Tested nginx timeout locally by calling the priming URL and
  confirming that it timed out and wasn't retried.

- Tested stderr logging locally by creating routes that write
  to stderr using console.warn and by throwing errors and confirmed
  that the warnings and errors are visible via docker logs.

Reviewers: csilvers, drosile

Reviewed By: csilvers, drosile

Differential Revision: https://phabricator.khanacademy.org/D40201