[ADR-28 3/3] Add a mechanism for fetching cacheable data
Summary:
This adds a `propsUrl` option to the render server.

`propsUrl` is a URL from which props can be fetched. If provided, a JSON GET request will be made to that URL, and the result will be passed to the component as props, instead of 'props'.

This can be used to enforce that props do not contain user-specific data (ADR #28), since the render server has no authentication.

Test Plan:
 - Added a test for `propsURL`
 - `npm run test`
 - Tested further in D43894.

Reviewers: briangenisio, russell

Reviewed By: russell

Spies: tom

Differential Revision: https://phabricator.khanacademy.org/D43901