Update target CPU use to 20%.
The react-render server is not scaling to handle the load of the video
page. With recent changes to the way we do SSR, it's bound on network
and/or memory, not CPU. In flex, we cannot sale based on the concurrent
connections or memory, so this sets the target CPU to about what it is
when the server is running as fast as it can.

https://khanacademy.slack.com/archives/C3UDL9QN7/p1537387282000100

Auditors: amos, benkraft