Get rid of support for propsUrl.
Summary:
This was a way we had of dealing with react-render-serving a component
that had user-specific data in it, without requiring the server to
actually know about the user-specific data.  But it never caught on,
and we've solved the problem by moving the user-specific logic out of
the initial render and into graphql calls.  So we can remove
propsUrl.  The client no longer uses it as of D49981.

This is just a partial revert of D43901.

Test Plan: npm test

Reviewers: joshuan

Reviewed By: joshuan

Subscribers: john, russell, briangenisio

Differential Revision: https://phabricator.khanacademy.org/D50094