Split out Apollo Network configuration
Summary: This is just a small refactor. Since we're going to be changing the render file I wanted to make it a little easier to follow, so this clearly distinct functionality has been moved to its own file. I have not added additional tests as yet, but we could if we wanted to.

Test Plan: `npm test`

Reviewers: john, joshuan

Reviewed By: joshuan

Differential Revision: https://phabricator.khanacademy.org/D51216