[WEB-543].3 A workflow that works with our webpack
Summary:
In this change, the entire workflow for rendering is shifted around. Instead of doing lots of work on the server, much of it is pushed to the client. So now, instead of the client being passive in this situation, it is responsible for initiating the render and doing the main rendering effort. The server sets up Apollo Client and handles the requests accordingly.

The new flow is this:

1. The server is given a list of file URLs to load
2. The server fetches and executes those URLs in order (from cache if it already got them)
3. The final URL loaded registers with server-side rendering (it can determine if this is needed by the existence of the __registerForSSR__ callback)
4. The server uses the callback provided to __registerForSSR__ to do the render, relying on the registered callback to do the work (including calling getDataForTree)
5. The server waits for the promise to resolve (expected result should be an object, {html, css})
6. On promise resolution, the server attaches the requested data and fulfills the original request with the results of the form

The data returned by the server is as it was before this change; its mainly the `/render` API and how it works that has changed.

Test Plan:
- `npm test`

Reviewers: john,kevinb,joshuan

Subscribers: #webpack