Up the instance size for app-standard
Summary:
With SSR live in webapp, we're seeing quite a few OOM errors (somewhere
around 5% of requests). The errors recommend upping the instance size
to cope with this. It was [suggested in a Slack thread](https://khanacademy.slack.com/archives/C3UDL9QN7/p1556649573112400?thread_ts=1556648461.110900&cid=C3UDL9QN7) that we should
follow the suggestion, so this change does that.

Test Plan:
Not sure what test we could do here. Suggestions are welcome.

Reviewers: csilvers,amos

Subscribers: #infrastructure, #fe-infra