Add profiling
Summary:
Adds the StackDriver profiling to the react-render-server per
https://cloud.google.com/profiler/docs/profiling-nodejs

Test Plan:
For dev:
Started up webapp, killed the react-render-server that it starts,
ran the react-render-server dev app with `npm start` and checked that
it works.

For prod:
I implemented this according to the docs. Not sure of the best way to
validate it before we actually deploy. Suggestions are welcome as I
think we need a better test story for changes like this.

Reviewers: csilvers,amos,#fe-infra

Subscribers: #infrastructure