Up the timeout on cache requests to 2 seconds
Summary:
We're seeing a lot of cache requests timing out in production. This
change ups that timeout to 2 seconds (from 1 second), and adds some
recording of the actual time a request takes (since the original
request is inflight still).

This way, we can see:

 1. Does the timeout at 2 seconds alleviate the error noise?
 2. What should the timeout be based on what the actual request times are?

This also puts the instance size back to `F4` as I don't think we need to
up that just yet (if at all).

Test Plan:
In dev there is no change here.

For production, we should deploy and monitor error rates and such.
If things look unfavourable, we can migrate traffic to the previous
version (this may require Infra/SRE support for permissions to migrate).

Reviewers: john, amos, csilvers

Subscribers: #frontend-infrastructure, #infrastructure

Differential Revision: https://phabricator.khanacademy.org/D54012